+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -115,7 +115,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed a warning in materialevalulation
 - Fixed an error when building the player.
 - Fixed issue with box light not visible if range is below one and range attenuation is off.
-- Fixed alpha not having TAA applied to it.
+- Fixed issue with bloom showing a thin black line after rescaling window. 
+- Fixed rendergraph motion vector resolve.
+- Fixed the Ray-Tracing related Debug Display not working in render graph mode.
+- Fix nan in pbr sky
 - Fixed Light skin not properly applied on the LookDev when switching from Dark Skin (case 1278802)
 
 ### Changed